Rename isolate components for clarity
