+++ mobile/apps/photos/lib/services/machine_learning/face_thumbnail_generator.dart
@@ -4,9 +4,9 @@ import 'dart:typed_data' show Uint8List;
 import "package:computer/computer.dart";
 import "package:logging/logging.dart";
 import "package:photos/models/ml/face/box.dart";
-import "package:photos/services/isolate_functions.dart";
-import "package:photos/services/isolate_service.dart";
 import "package:photos/utils/image_ml_util.dart";
+import "package:photos/utils/isolate/isolate_operations.dart";
+import "package:photos/utils/isolate/super_isolate.dart";
 
 final Computer _computer = Computer.shared();
